refactor: send Banner's data through props from App.jsx
