+++ src/App.jsx
@@ -14,6 +14,14 @@ import Footer from "./Components/Footer/Footer";
 import Cart from "./Pages/Cart";
 import Orders from "./Pages/Orders";
 
+// Banner Images
+import imgL1 from "./Assets/Banner_images/imgL1.png"
+import imgL2 from "./Assets/Banner_images/imgL2.png"
+import imgL3 from "./Assets/Banner_images/imgL3.png"
+import imgR1 from "./Assets/Banner_images/imgR1.png"
+import imgR2 from "./Assets/Banner_images/imgR2.png"
+import imgR3 from "./Assets/Banner_images/imgR3.png"
+
 const App = () => {
   const [loading, setLoading] = useState(true)
   const dispatch = useDispatch()
@@ -48,10 +56,10 @@ const App = () => {
   }, [dispatch, cartData.data, cartData.isError])
 
   useEffect(() => {
-    if(!orderData.isError && orderData.isSuccess) {
+    if (!orderData.isError && orderData.isSuccess) {
       dispatch(AddAllOrders(orderData.data))
     }
-  },[dispatch, orderData.data, orderData.data])
+  }, [dispatch, orderData.data, orderData.data])
 
 
   let categoryPath;
@@ -88,8 +96,61 @@ const App = () => {
     }
   }
 
-  let content
+  const leftBanner_Data = [
+    {
+      id: 1,
+      link: imgL1,
+      price:"499.99",
+      name: "Meta Quest 3",
+      title: "Meta Quest 3",
+      description: "Meta Oculus Quest 3 Advanced All-in-One Virtual Reality Headset (128GB)",
+    },
+    {
+      id: 2,
+      link: imgL2,
+      price:"199.00",
+      name: "Rabbit R1",
+      title: "R1. your pocket companion",
+      description: "Rabbit R1, your pocket companion",
+    },
+    {
+      id: 3,
+      link: imgL3,
+      price:"1300.00",
+      name: "Samsung Galaxy S24 Ultra",
+      title: "Samsung Galaxy S24 Ultra",
+      description: "Samsung Galaxy S24 Ultra 5G AI Smartphone",
+    },
+  ]
 
+  const rightBanner_Data = [
+    {
+      id: 4,
+      link: imgR1,
+      price:"1359.99",
+      name: "TP-7",
+      title: "TP-7 field recorder",
+      description: "Intuitive muscle reflex control"
+    },
+    {
+      id: 5,
+      link: imgR2,
+      price:"1299.99",
+      name: "Galaxy Tab S9 Ultra",
+      title: "Galaxy Tab S9 Ultra",
+      description: "Samsung Galaxy Tab S9 Ultra 36.99 cm (14.6 inch) Dynamic AMOLED 2X Display, RAM 12 GB, ROM 512 GB Expandable, S Pen in-Box, Wi-Fi Tablet, Gray"
+    },
+    {
+      id: 6,
+      link: imgR3,
+      price:"1799.99",
+      name: "MacBook Pro M3 Pro",
+      title: "Apple the new MacBook Pro featuring the M3 family of chips, making the world's best pro laptop even better",
+      description: "Apple 2023 MacBook Pro (14-inch, M3 Pro chip with 11-core CPU and 14-core GPU, 18GB Unified Memory, 512GB) - Space Grey"
+    },
+  ]
+
+  let content
   if (loading) {
     content = (
       <div className="w-full h-full flex flex-row justify-center items-center">
@@ -115,10 +176,10 @@ const App = () => {
   } else if (!loading && productsData.isSuccess) {
     content = (
       <div>
-        <Route path='/'><Home /></Route>
+        <Route path='/'><Home leftBanner_Data={leftBanner_Data} rightBanner_Data={rightBanner_Data} /></Route>
         <Route path={categoryPath}><ShowProducts category={categories[cateIndex]} /></Route>
         <Route path={productPath}><ProductDetail product={productData} /></Route>
-        <Route path='/cart'><Cart/></Route>
+        <Route path='/cart'><Cart /></Route>
         <Route path='/orders'><Orders /></Route>
       </div>
     )
@@ -129,7 +190,7 @@ const App = () => {
         <NavBar />
         {content}
         <div className="shadow-full shadow-indigo-500/50 mt-20">
-          <hr className="border-opacity-70 border-gray-700"/>
+          <hr className="border-opacity-70 border-gray-700" />
         </div>
         <Footer />
       </div>
