+++ src/components/Banners/BannerSlider.jsx
@@ -45,8 +45,8 @@ const BannerSlider = ({ data }) => {
                     </div>
                 </div>
                 <img className="select-none border-none outline-none rounded-3xl" src={product.link} alt={product.name} />
-                <Link 
-                    to={product.route}
+                <Link
+                    to={`/gadgets/${product.id}`}
                     key={product.id}
                     className={`flex flex-row justify-between items-center absolute bottom-8 right-16 p-1 pr-2 pl-2 text-base rounded-xl font-normal text-white bg-gray-900 bg-opacity-50 outline-none shadow-full shadow-gray-100/50 hover:bg-gray-100 hover:text-gray-900 hover:shadow-full hover:shadow-gray-900`}
                 >
